Action buttons visibility depending on permissions

+++ app/views/page_list.php
@@ -6,7 +6,9 @@
             <th>Auteur</th>
             <th>Statut</th>
             <th>Créé le</th>
-            <th></th>
+            <?php if ($can_delete) : ?>
+                <th></th>
+            <?php endif; ?>
         </tr>
         </thead>
         <tbody>
@@ -14,7 +16,7 @@
             <tr>
                 <td>
                     <p class="flex-col justify-between items-center">
-                        <a href="<?= $page['url_detail'] ?>"><?= $page['title'] ?></a>
+                        <a <?= $can_read ? 'href="' . $page['url_detail'] . '"' : '' ?>><?= $page['title'] ?></a>
                         <?php if ($page['status'] == \App\Core\Utils\Constants::STATUS_PUBLISHED) : ?>
                             <em style="font-size:.75em"><?= $page['slug'] ?></em>
                         <?php endif; ?>
@@ -23,11 +25,13 @@
                 <td><?= $page['author'] ?></td>
                 <td><?= $page['status_label'] ?></td>
                 <td><?= $page['created_at'] ?></td>
-                <td>
-                    <a href="<?= $page['url_delete'] ?>" data-role="deleteItem">
-                        <i class="fas fa-trash-alt"></i>
-                    </a>
-                </td>
+                <?php if ($can_delete) : ?>
+                    <td>
+                        <a href="<?= $page['url_delete'] ?>" data-role="deleteItem">
+                            <i class="fas fa-trash-alt"></i>
+                        </a>
+                    </td>
+                <?php endif; ?>
             </tr>
         <?php endforeach; ?>
         </tbody>
